Made authentication in one file each
